requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
cohere==3.2.3
courlan==0.8.3
Cython==0.29.33
dateparser==1.1.6
demucs==4.0.0
diffq==0.2.3
dnspython==2.3.0
dora-search==0.1.11
einops==0.6.0
fastapi==0.89.1
frozenlist==1.3.3
h11==0.14.0
htmldate==1.4.1
idna==3.4
julius==0.2.7
jusText==3.0.0
lameenc==1.4.2
langcodes==3.3.0
loguru==0.6.0
lxml==4.9.2
multidict==6.0.4
numpy==1.24.1
omegaconf==2.3.0
openai==0.26.4
openunmix==1.2.1
pinecone-client==2.1.0
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
retrying==1.3.4
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
submitit==1.4.5
tld==0.12.6
torch==1.13.1
torchaudio==0.13.1
tqdm==4.64.1
treetable==0.2.5
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
yarl==1.8.2
